+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Change shard computation to be based on number of head series.
+
 ## [4.42.0] - 2023-06-26
 
 ### Added